requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
dependency-injector==4.41.0
Django==4.2.8
djangorestframework==3.14.0
idna==3.4
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.6
django-filter~=23.5
